feat: Improving project structure, separating api calls from pages, minor improvements
