+++ frontend-react/src/constants/commonFilters.js
@@ -1,0 +1,15 @@
+export const booleanList = [{
+    name: "Yes",
+    value: true
+}, {
+    name: "No",
+    value: false
+}];
+
+export const sortOrdersList = [{
+    name: "Ascending",
+    value: "ASC"
+}, {
+    name: "Descending",
+    value: "DESC"
+}]
